Fixed problem with existing S3 methods
When an existing method was an S3 generic, the mock method would create
two generics so that the default method would infinitely recurse by
calling itself. A check was added and function only assigned as S3 
generic if it wasn't already one. Tests were also added to test this
functionality. This test also used existing plot method, which caused 
problem of being able to write to functions namespace. This was changed
to write to mockR package environment instead.
Idea for future is to create all mock methods in their own environment
on the search path so we don't have to mess with existing functions.
This will only work if this env can be searched in priority as needed.