Trading multiple securities
Somehow a transaction was being written with zero values, which was
introducing a NaN value for cash and holdings. Did not find the source
of the zero values, but resolved by making account updates handle the 
zero entry (in calculate_missing_prices). Test was added to Account
to reflect this.
Have tested successfully on up to three instruments (AMP, BHP, CBA).
At present the very simple strategy is causing positions to be taken
with more value than current equity.
Next steps to sort out stop order placement.