Backtesting with stops almost there
New order class to submit a stop order on the day that a market order
is executed. This is to simulate setting a stop manually after the open.
Part way through adding a position status object to manage the behaviour
of positions depending on what state they are in.
ID for orders was changed to a unique character string as orders were
being added with duplicate IDs when an order remained in the order book.